Move to poetry, move to GHA, deprecate EOL Python version support
