--- .circleci/config.yml
@@ -1,66 +1,0 @@
-version: 2
-workflows:
-  version: 2
-  test:
-    jobs:
-      - test-3.6
-      - test-3.7
-      - test-3.8
-      - test-3.9
-      - test-3.10
-      - test-3.11
-jobs:
-  test-3.6: &test-template
-    filters:
-      branches:
-        ignore: master
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.6
-    docker:
-      - image: cimg/python:3.6
-    steps:
-      - checkout
-      - run:
-          name: install dependencies
-          command: |
-            pip install --user pytest>=4.6.5 more-itertools>=5.0.0 pathlib2 configparser
-            pip install --user -e .
-      - run:
-          name: run tests
-          command: |
-            python -m pytest
-
-  test-3.7:
-    <<: *test-template
-    docker:
-      - image: cimg/python:3.7
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.7
-
-  test-3.8:
-    <<: *test-template
-    docker:
-      - image: cimg/python:3.8
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.8
-
-  test-3.9:
-    <<: *test-template
-    docker:
-      - image: cimg/python:3.9
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.9
-
-  test-3.10:
-    <<: *test-template
-    docker:
-      - image: cimg/python:3.10
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.10
-
-  test-3.11:
-    <<: *test-template
-    docker:
-      - image: cimg/python:3.11
-    environment:
-      TRAVIS_PYTHON_VERSION: 3.11
