+++ .github/workflows/python-version-tests.yml
@@ -1,0 +1,35 @@
+name: Python Version Tests
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.10", "3.11", "3.12", "3.13", "3.14"]
+
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install Poetry and Dependencies
+      run: |
+        make quickstart 
+    - name: Run lint checks
+      run: |
+        make lint
+    - name: Run tests
+      run: |
+        make test
