+++ doc/make.bat
@@ -39,8 +39,8 @@ if "%1" == "help" (
 )
 
 if "%1" == "clean" (
-	for /d %%i in (%BUILDDIR%\*) do rmdir /q /s %%i
-	del /q /s %BUILDDIR%\*
+	for /d %%i in (%BUILDDIR%\\*) do rmdir /q /s %%i
+	del /q /s %BUILDDIR%\\*
 	goto end
 )
