+++ setup.cfg
@@ -1,0 +1,5 @@
+[wheel]
+universal = 1
+
+[flake8]
+max-line-length = 119
